requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
openai==1.3.7
azure-identity==1.15.0
langchain==0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3
aiofiles==23.2.1
jinja2==3.1.2
huggingface-hub==0.20.3
sentence-transformers==2.7.0 